On 32 bits, SQLINTEGER is probably defined as long, so gcc complains.

+++ clients/odbc/driver/ODBCUtil.c
@@ -1277,7 +1277,7 @@ translateConnectAttribute(SQLINTEGER Attribute)
 	case SQL_ATTR_TXN_ISOLATION:
 		return "SQL_ATTR_TXN_ISOLATION";
 	default:
-		snprintf(unknown, sizeof(unknown), "unknown (%d)", Attribute);
+		snprintf(unknown, sizeof(unknown), "unknown (%d)", (int) Attribute);
 		return unknown;
 	}
 }
@@ -1328,7 +1328,7 @@ translateEnvAttribute(SQLINTEGER Attribute)
 	case SQL_ATTR_CP_MATCH:
 		return "SQL_ATTR_CP_MATCH";
 	default:
-		snprintf(unknown, sizeof(unknown), "unknown (%d)", Attribute);
+		snprintf(unknown, sizeof(unknown), "unknown (%d)", (int) Attribute);
 		return unknown;
 	}
 }
@@ -1406,7 +1406,7 @@ translateStmtAttribute(SQLINTEGER Attribute)
 	case SQL_ATTR_USE_BOOKMARKS:
 		return "SQL_ATTR_USE_BOOKMARKS";
 	default:
-		snprintf(unknown, sizeof(unknown), "unknown (%d)", Attribute);
+		snprintf(unknown, sizeof(unknown), "unknown (%d)", (int) Attribute);
 		return unknown;
 	}
 }
